Move tenant-aware integration tests to separate class (#405)
Move tenant-aware integration tests to separate class. This simplifies the setup and teardown required for these tests.
